CM-998 unit tests optimization and pytest-xdist support added. Run pytest in parallel with `pytest -n auto`
