+++ .gitignore
@@ -59,6 +59,7 @@ docs/_build/
 *~
 *.pyc
  /.idea/
+ .idea
 
 #updated ontology files
 Climate_Mind_Digraph_Test_Ont.json
