+++ app/conftest.py
@@ -1,19 +1,6 @@
 import factory
 import pytest
 from flask_sqlalchemy import SQLAlchemy
-from sqlalchemy import create_engine
-from sqlalchemy.exc import ProgrammingError
-
-from app import create_app
-from app.extensions import db
-from config import DevelopmentConfig, TestingConfig
-
-
-def pytest_configure():
-    """Make the following constants available across all pytest unit tests."""
-    pytest.config = {
-        "TEST_PASSWORD": "reduce_dairy_and_meat_consumption_666",
-    }
 
 
 @pytest.fixture(autouse=True)
@@ -25,47 +12,9 @@ def set_session_for_factories(db_session):
     yield
 
 
-class AuthActions(object):
-    def __init__(self, client):
-        self._client = client
-
-    def login(self, username, password):
-        return self._client.post(
-            "/auth/login", data={"username": username, "password": password}
-        )
-
-    def logout(self):
-        return self._client.get("/auth/logout")
-
-
-@pytest.fixture
-def auth(client):
-    return AuthActions(client)
-
-
 @pytest.fixture(scope="session")
-def database_instance():
-    with create_engine(
-        DevelopmentConfig.SQLALCHEMY_DATABASE_URI,
-        isolation_level="AUTOCOMMIT",
-        echo=True,
-    ).connect() as connection:
-        try:
-            connection.execute(f"CREATE DATABASE [{TestingConfig.DB_NAME}]")
-        except ProgrammingError:
-            connection.execute(f"DROP DATABASE [{TestingConfig.DB_NAME}]")
-            connection.execute(f"CREATE DATABASE [{TestingConfig.DB_NAME}]")
-
-        yield
-        connection.execute(f"DROP DATABASE [{TestingConfig.DB_NAME}]")
-
-
-@pytest.fixture(scope="session")
-def app(database_instance):
-    app = create_app(config_class=TestingConfig)
-    with app.app_context():
-        db.create_all()
-        yield app
+def app(worker_id):
+    yield pytest.config["APP"]
 
 
 @pytest.fixture(scope="session")
