+++ config.py
@@ -1,4 +1,3 @@
-from flask import abort
 import os
 import urllib
 from datetime import timedelta
